pyproject.toml
[tool.poetry]
name = "paperqa"
version = "0.1.0"
description = ""
authors = ["Eike Kohlmeyer <ekohlmeyer21@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.7"
streamlit = "^1.22.0"
cloudpathlib = "^0.14.0"
boto3 = "^1.26.138"
farm-haystack = {extras=["faiss,colab"], version="^1.17.0"}

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
flake8 = "^6.0.0"
black = "^23.3.0"
pre-commit = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
